# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "x402 API Portfolio",
    "version": "1.0.0",
    "description": "Paid REST APIs on Base using x402 and USDC. A 402 response advertises payment requirements; send the signed payload in X-PAYMENT and retry."
  },
  "servers": [
    {
      "url": "https://x402-api-91r3.vercel.app",
      "description": "Production"
    }
  ],
  "security": [
    {
      "x402Payment": []
    }
  ],
  "paths": {
...
    "/api/scraped-data": {
      "get": {
        "operationId": "scrapeData",
        "summary": "Scrape a target URL",
        "parameters": [
          {
            "name": "target",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uri",
              "default": "https://example.com"
            },
            "description": "Absolute HTTP or HTTPS URL to scrape. Defaults to https://example.com."
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/ScrapeResponse",
            "headers": {
              "PAYMENT-RESPONSE": {
                "$ref": "#/components/headers/PaymentResponse"
              }
            }
          },
          "400": {
            "description": "Invalid target URL.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Invalid target URL"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "402": {
            "$ref": "#/components/responses/PaymentRequired"
          },
          "500": {
            "description": "Unexpected server error."
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "x402Payment": {
        "type": "apiKey",
        "in": "header",
        "name": "X-PAYMENT",
        "description": "Base64-encoded x402 payment payload sent on the retry request."
      }
    },
    "headers": {
      "PaymentRequired": {
        "description": "Base64-encoded x402 payment requirements. Includes network, USDC asset, payTo, resource, and amount.",
        "schema": {
          "type": "string"
        }
      },
      "PaymentResponse": {
        "description": "Base64-encoded x402 settlement response, including the settlement transaction hash when available.",
        "schema": {
          "type": "string"
        }
      }
    },
...
    "responses": {
      "PaymentRequired": {
        "description": "HTTP 402 payment challenge. The requirements identify the USDC asset, required amount in six-decimal atomic units, Base network, and payTo facilitator receiver address. Sign the EIP-3009 authorization, send it as X-PAYMENT, and retry. Successful responses include PAYMENT-RESPONSE.",
        "headers": {
          "PAYMENT-REQUIRED": {
            "$ref": "#/components/headers/PaymentRequired"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/PaymentChallenge"
            }
          }
        }
      },
...
      "ScrapeResponse": {
        "description": "Paid scraped data",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ScrapeResult"
            }
          }
        }
      }
    },
    "schemas": {
      "EvmAddress": {
        "type": "string",
        "pattern": "^0x[a-fA-F0-9]{40}$",
        "example": "0x0000000000000000000000000000000000000001"
      },
      "PaymentChallenge": {
        "type": "object",
        "required": [
          "x402Version",
          "accepts"
        ],
        "properties": {
          "x402Version": {
            "type": "integer",
            "example": 1
          },
          "accepts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentRequirement"
            }
          }
        }
      },
      "PaymentRequirement": {
        "type": "object",
        "required": [
          "scheme",
          "network",
          "asset",
          "payTo",
          "maxAmountRequired",
          "resource"
        ],
        "properties": {
          "scheme": {
            "type": "string",
            "example": "exact"
          },
          "network": {
            "type": "string",
            "example": "base"
          },
          "asset": {
            "type": "string",
            "example": "USDC"
          },
          "payTo": {
            "$ref": "#/components/schemas/EvmAddress"
          },
          "maxAmountRequired": {
            "type": "string",
            "description": "USDC atomic units with 6 decimals"
          },
          "resource": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "ScrapeResult": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          }
        }
      }
    }
  }
}